fixed AOCS misunderstanding
previously, the norm of the control thurst was computed according to the current position of the spacecraft. Instead now the control thurst norm is computed from the position in which the spacecraft is supposed to be. This means that the thrust in the trajectory control and attitude control cases will be the same just oriented differently.