fixed signvect issue
fixed an issue which occurred when the matrix2Quat function was called in the regenerative trajectory. The routine by itself updates every time the signvect variable, so the signvect would be updated to match the signs at the end of the commandend attitude propagation rather than at the current time, leading to a jump in commanded quaternions and therefore a jump in error quaternions.